pyproject.toml
[project]
name = "gts-graph-rag"
version = "0.1.0"
description = "HSBC BOI Agentic GraphRAG System for knowledge transfer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # LangGraph & LangChain
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    # Neo4j GraphRAG
    "neo4j>=5.23.0",
    "neo4j-graphrag>=1.0.0",
    # Vector Store
    "chromadb>=0.5.0",
    # LLM & Embeddings (DashScope)
    "openai>=1.50.0",
    "dashscope>=1.20.0",
    # FastAPI Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "sse-starlette>=2.1.0",
    # Document Processing
    "pypdf>=5.0.0",
    "openpyxl>=3.1.2",
    "python-docx>=1.1.0",
    "unstructured>=0.15.0",
    # Configuration
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    # Auth & Database
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "sqlalchemy>=2.0.0",
    "bcrypt==4.0.1",
    "pymupdf>=1.26.7",
    "langchain-openai>=1.1.6",
    "langchain-experimental>=0.4.1",
    "langchain-neo4j>=0.6.0",
    # URL Parsing for temp knowledge
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.27.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

# =============================================================================
# Ruff Configuration (PEP 8 Linting)
# =============================================================================
[tool.ruff]
line-length = 79
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "N",      # pep8-naming
]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Function call in default argument (FastAPI Depends)
    "N806",   # Variable in function should be lowercase
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "B"]  # Ignore docstring and bugbear rules in tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# =============================================================================
# Mypy Configuration (Type Checking)
# =============================================================================
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# =============================================================================
# Pytest Configuration
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

# =============================================================================
# Coverage Configuration
# =============================================================================
[tool.coverage.run]
source = ["backend"]
omit = ["tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
